backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.6
django-core==1.4.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
pipenv==2020.6.2
pylint==2.5.2
pytz==2020.1
requests==2.24.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
